+++ bundle/src/Testing/ApiTestingTrait.php
@@ -2,6 +2,7 @@
 
 namespace Hyvor\Internal\Bundle\Testing;
 
+use Symfony\Component\BrowserKit\AbstractBrowser;
 use Symfony\Component\HttpFoundation\Response;
 
 trait ApiTestingTrait
@@ -14,7 +15,9 @@ trait ApiTestingTrait
      */
     public function getJson(): array
     {
-        $response = self::getClient()->getResponse();
+        /** @var AbstractBrowser<object, \Symfony\Component\BrowserKit\Response> $client */
+        $client = self::getClient();
+        $response = $client->getResponse();
         $content = $response->getContent();
         $this->assertNotFalse($content);
         $this->assertJson($content);
@@ -23,9 +26,19 @@ trait ApiTestingTrait
         return $json;
     }
 
+    /**
+     * @deprecated use assertResponseFailed
+     */
     public function assertFailed(int $code, ?string $message = null): void
     {
-        $response = self::getClient()->getResponse();
+        $this->assertResponseFailed($code, $message);
+    }
+
+    public function assertResponseFailed(int $code, ?string $message = null): void
+    {
+        /** @var AbstractBrowser<object, \Symfony\Component\BrowserKit\Response> $client */
+        $client = self::getClient();
+        $response = $client->getResponse();
         $this->assertSame($code, $response->getStatusCode());
 
         if ($message !== null) {
@@ -36,7 +49,9 @@ trait ApiTestingTrait
 
     public function assertViolationCount(int $count): void
     {
-        $response = self::getClient()->getResponse();
+        /** @var AbstractBrowser<object, \Symfony\Component\BrowserKit\Response> $client */
+        $client = self::getClient();
+        $response = $client->getResponse();
         $this->assertSame(Response::HTTP_UNPROCESSABLE_ENTITY, $response->getStatusCode());
 
         $response = $this->getJson();
@@ -48,7 +63,9 @@ trait ApiTestingTrait
 
     public function assertHasViolation(string $property, string $message = ''): void
     {
-        $response = self::getClient()->getResponse();
+        /** @var AbstractBrowser<object, \Symfony\Component\BrowserKit\Response> $client */
+        $client = self::getClient();
+        $response = $client->getResponse();
         $this->assertSame(Response::HTTP_UNPROCESSABLE_ENTITY, $response->getStatusCode());
 
         $response = $this->getJson();